Add authentication and voting methods to GbxClient; refactor session and script handling
