+++ gbxclient/gbxclient.go
@@ -1,4 +1,4 @@
-package main
+package gbxclient
 
 import (
 	"bytes"
