+++ gbxclient/methods.script.go
@@ -1,4 +1,4 @@
-package main
+package gbxclient
 
 import (
 	"errors"
